% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{count_matches}
\alias{count_matches}
\title{Count regex pattern matches}
\usage{
count_matches(pattern, value)
}
\arguments{
\item{pattern}{Regular expression pattern}

\item{value}{String to search in}
}
\value{
Number of matches found
}
\description{
Count regex pattern matches
}
\keyword{internal}
