% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{\%||\%}
\alias{\%||\%}
\title{NULL coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value}

\item{b}{Default value if a is NULL}
}
\value{
a if not NULL, otherwise b
}
\description{
NULL coalescing operator
}
