% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_image.R
\name{image_validator}
\alias{image_validator}
\title{Image Validator}
\usage{
image_validator(
  file,
  max_size = Inf,
  allowed_types = c("image/png", "image/jpeg", "image/jpg", "image/gif"),
  max_width = Inf,
  max_height = Inf,
  message = NULL
)
}
\arguments{
\item{file}{The file to validate (from Shiny's \code{input$file}).}

\item{max_size}{Maximum file size in bytes. Default is Inf.}

\item{allowed_types}{Allowed image types. Default is c("image/png", "image/jpeg", "image/jpg", "image/gif").}

\item{max_width}{Maximum width in pixels. Default is Inf.}

\item{max_height}{Maximum height in pixels. Default is Inf.}

\item{message}{Custom error message. If NULL, default messages are used.}
}
\value{
A list with \code{is_valid} (logical) and \code{message} (character).
}
\description{
Validates that a file is an image and meets specified criteria.
}
\examples{
# In a Shiny app, use image_validator(input$file)
}
