% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_not_blank.R
\name{not_blank}
\alias{not_blank}
\title{NotBlank Validator}
\usage{
not_blank(value, message = NULL)
}
\arguments{
\item{value}{The value to validate.}

\item{message}{Custom error message. If NULL, a default message is used.}
}
\value{
A list with \code{is_valid} (logical) and \code{message} (character).
}
\description{
Validates that the input is not NULL or an empty string.
}
\examples{
not_blank("Hello")
not_blank("")
}
