% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_not_compromised_password.R
\name{not_compromised_password}
\alias{not_compromised_password}
\title{Not Compromised Password Validator}
\usage{
not_compromised_password(value, message = NULL)
}
\arguments{
\item{value}{The password to validate.}

\item{message}{Custom error message. If NULL, a default message is used.}
}
\value{
A list with \code{is_valid} (logical) and \code{message} (character).
}
\description{
Validates that a password has not been compromised in known data breaches.
Uses the Have I Been Pwned API.
}
\note{
This function uses the Have I Been Pwned API and the k-anonymity model.
}
\examples{
not_compromised_password("P@ssw0rd!")
}
