% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_password_strength.R
\name{password_strength}
\alias{password_strength}
\title{Password Strength Validator}
\usage{
password_strength(
  value,
  min_length = 8,
  min_upper = 1,
  min_lower = 1,
  min_digit = 1,
  min_special = 1,
  custom_message = NULL
)
}
\arguments{
\item{value}{The password to validate.}

\item{min_length}{Minimum length. Default is 8.}

\item{min_upper}{Minimum number of uppercase letters. Default is 1.}

\item{min_lower}{Minimum number of lowercase letters. Default is 1.}

\item{min_digit}{Minimum number of digits. Default is 1.}

\item{min_special}{Minimum number of special characters. Default is 1.}

\item{custom_message}{Custom error message. If NULL, default messages are used.}
}
\value{
A list with \code{is_valid} (logical) and \code{message} (character).
}
\description{
Validates that a password meets strength requirements.
}
\examples{
password_strength("P@ssw0rd!")
password_strength("weakpassword", custom_message = "Password is too weak.")
}
