% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvalidator-package.R
\docType{package}
\name{rvalidator-package}
\alias{rvalidator}
\alias{rvalidator-package}
\title{rvalidator: Validates Values Based On Constraints}
\description{
Validates values based on built-in and custom onstraints. Provides also integration for use in Shiny apps.
}
\author{
\strong{Maintainer}: Sasha Bogdanovic \email{sasha@cityviz.ca}

}
\keyword{internal}
