% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validate_input.R
\name{validate_input_server}
\alias{validate_input_server}
\title{Validate Input Module Server}
\usage{
validate_input_server(id, validators)
}
\arguments{
\item{id}{The module id.}

\item{validators}{A list of validator functions.}
}
\value{
Reactive value of the input.
}
\description{
Validate Input Module Server
}
