% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validate_input.R
\name{validate_input_ui}
\alias{validate_input_ui}
\title{Validate Input Module UI}
\usage{
validate_input_ui(id, input_type = "textInput", label, ...)
}
\arguments{
\item{id}{The module id.}

\item{input_type}{The type of input (e.g., "textInput", "numericInput").}

\item{label}{The label of the input.}

\item{...}{Additional arguments passed to the input function.}
}
\value{
Shiny UI element.
}
\description{
Validate Input Module UI
}
